fdgd

+++ database/migrations/2022_05_21_183124_create_ichkitavars_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->bigInteger('tavar2_id')->unsigned();
             $table->string('name')->nullable();
             $table->string('raqam')->nullable();
-            $table->integer('hajm')->nullable();
+            $table->float('hajm')->nullable();
             $table->integer('summa')->nullable();
             $table->integer('summa2')->nullable();
             $table->integer('summa3')->nullable();
